+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/TestReceiver.kt
@@ -1,0 +1,12 @@
+package me.panpf.androidxkt.test
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class TestReceiver : BroadcastReceiver() {
+
+    override fun onReceive(context: Context, intent: Intent) {
+
+    }
+}