+++ androidx/src/androidTest/java/me/panpf/androidx/test/TestReceiver.java
@@ -1,0 +1,13 @@
+package me.panpf.androidx.test;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+public class TestReceiver extends BroadcastReceiver {
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+
+    }
+}